server/kivijalka.sh: Make sure there's an image for qiv to read when it's started.

+++ server/kivijalka.sh
@@ -47,6 +47,12 @@ cutycapt --url="$URL" \
   --min-height=$PX_H \
   --out="$OUTPATH/$CAP"
 
+if [[ ! -f "$OUTPATH/$CAP" ]]; then
+  exit 1
+fi
+
+cp "$OUTPATH/$CAP" "$OUTPATH/$OUT"
+
 (
   echo $BASHPID > "$PIDFILE"
   while true
